requirements2.txt
Flask
Flask-cors
weaviate-client
langchain
sentence-transformers
tiktoken
pypdf
rapidocr-onnxruntime
huggingface-hub
